+++ packages/discord.js/src/structures/ChatInputCommandInteraction.js
@@ -2,6 +2,7 @@
 
 const CommandInteraction = require('./CommandInteraction');
 const CommandInteractionOptionResolver = require('./CommandInteractionOptionResolver');
+const { transformResolved } = require('../util/transformResolved');
 
 /**
  * Represents a command interaction.
@@ -18,7 +19,7 @@ class ChatInputCommandInteraction extends CommandInteraction {
     this.options = new CommandInteractionOptionResolver(
       this.client,
       data.data.options?.map(option => this.transformOption(option, data.data.resolved)) ?? [],
-      this.transformResolved(data.data.resolved ?? {}),
+      transformResolved({ client: this.client, guild: this.guild, channel: this.channel }, data.data.resolved),
     );
   }
